+++ src/main/java/br/com/desafio/repository/QuartoRepository.java
@@ -1,6 +1,5 @@
 package br.com.desafio.repository;
 
-import br.com.desafio.model.Hotel;
 import br.com.desafio.model.Quarto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -8,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface QuartoRepository extends JpaRepository<Quarto, Integer> {
+public interface QuartoRepository extends JpaRepository<Quarto, Long> {
     Quarto findByNumQuarto(int numQuarto);
     List<Quarto> findByResevadoFalse();
     List<Quarto> findByResevadoTrue();
